Improve ::selection documentation which became confusing after c7057be
After Autoprefixer was added a comment explaining that the vendor prefixed
`::-moz-selection` and the default `::selection` can't be combined became
obsolete. I rephrased the comment to be more general and so that it still
makes sense in the dist css file. A stackoverflow link explains why vendor
prefixed selectors can't be combined with other selectors.

Fixes #1953